Improve email service by spliting the providers into different classes (#236)
First, just wanted to say that I really like the project and the work
you've done so far!

This PR is mostly a minor improvement on how we are capturing the email
provider information -- main reason because i wanted to add
https://resend.com/ but didn't want to make the if block larger than it
is.

This PR:

- Creates an abstract class that mailers can extend. Main reason to do
it this way is to ensure consistent typings/returns and hide those
details on each mailer that implements it
- Creates a mailer class for each, that way we can isolate any future
nuances and eventually add more providers (e.g. resend, in another PR)
👌🏽


Feel free to discard or take over the PR if you don't like the approach
of classes, I'm not married to it.